Fixes Issue #15: Removes "do_cloudflare" checks. No credentials for /ips endpoint. (PR #17)
Per our discussion here https://github.com/wp-media/cloudflare/issues/15#issuecomment-588290706, this PR does the following:

1. Closes #15 
2. Removes `do_cloudflare` checks and code.
3. Makes `/ips` endpoint not need API credentials.
