+++ Tests/Integration/Subscriber/autoPurgeByUrl.php
@@ -22,15 +22,6 @@ class Test_AutoPurgeByUrl extends TestCase {
 		self::$post_id = $factory->post->create();
 	}
 
-	public function testShouldBailoutWhenCFAddonOff() {
-		$this->setOptions( [ 'do_cloudflare' => 0 ] );
-
-		Functions\expect( 'current_user_can' )->with( 'rocket_purge_cloudflare_cache' )->never();
-		Functions\expect( 'get_rocket_i18n_home_url' )->never();
-
-		do_action( 'after_rocket_clean_post', self::$post_id, [], 'en' );
-	}
-
 	public function testShouldBailoutWhenUserCantPurgeCF() {
 		$user = $this->factory->user->create( [ 'role' => 'contributor' ] );
 		wp_set_current_user( $user );
@@ -52,10 +43,5 @@ class Test_AutoPurgeByUrl extends TestCase {
 		Functions\expect( 'get_rocket_i18n_home_url' )->never();
 
 		do_action( 'after_rocket_clean_post', self::$post_id, [], 'en' );
-	}
-
-	private function setOptions( $options ) {
-		update_option( 'wp_rocket_settings', $options );
-		self::$options->set_values( $options );
 	}
 }
