--- Tests/Unit/CloudflareSubscriber/purgeCache.php
@@ -1,163 +1,0 @@
-<?php
-namespace WPMedia\Cloudflare\Tests\Unit\CloudflareSubscriber;
-
-use Brain\Monkey\Functions;
-use Mockery;
-use WPMedia\Cloudflare\Cloudflare;
-use WPMedia\Cloudflare\Subscriber;
-use WPMedia\Cloudflare\Tests\Unit\TestCase;
-
-/**
- * @covers WPMedia\Cloudflare\CloudflareSubscriber::purge_cache
- * @group  Cloudflare
- * @group  CloudflareSubscriber
- */
-class Test_PurgeCache extends TestCase {
-
-	/**
-	 * Test should not Purge Cloudflare if Cloudflare addon is disabled.
-	 */
-	public function testShouldNotPurgeWhenCloudflareIsDisabled() {
-		$mocks = $this->getConstructorMocks( 0, '', '', '' );
-
-		$cloudflare = Mockery::mock( Cloudflare::class);
-		$cloudflare->shouldNotReceive('purge_cloudflare');
-
-		$_GET['_wpnonce'] = 'nonce';
-		Functions\when( 'wp_verify_nonce' )->justReturn( true );
-		Functions\when( 'sanitize_key' )->justReturn( '' );
-
-		$cloudflare_subscriber = new Subscriber( $cloudflare, $mocks['options_data'], $mocks['options'] );
-		$cloudflare_subscriber->purge_cache_no_die();
-	}
-
-	/**
-	 * Test should not Purge Cloudflare if Cloudflare addon is enabled but user has no permissions.
-	 */
-	public function testShouldNotPurgeWhenCloudflareIsEnabledNoUserPermission() {
-		$mocks = $this->getConstructorMocks( 1, '', '', '' );
-
-		$cloudflare = Mockery::mock( Cloudflare::class);
-		$cloudflare->shouldNotReceive('purge_cloudflare');
-
-		$_GET['_wpnonce'] = '';
-		Functions\when( 'wp_verify_nonce' )->justReturn( true );
-		Functions\when( 'sanitize_key' )->justReturn( '' );
-		Functions\when( 'current_user_can' )->justReturn( false );
-
-		$cloudflare_subscriber = new Subscriber( $cloudflare, $mocks['options_data'], $mocks['options'] );
-		$cloudflare_subscriber->purge_cache_no_die();
-	}
-
-	/**
-	 * Test should Purge Cloudflare and return error.
-	 */
-	public function testShouldPurgeWithError() {
-		$mocks = $this->getConstructorMocks( 1, '', '', '' );
-
-		Functions\when( 'is_wp_error' )->justReturn( true );
-
-		$wp_error   = \Mockery::mock( \WP_Error::class );
-		$wp_error->shouldReceive('get_error_message')->andReturn( 'Error!' );
-
-		$cloudflare = Mockery::mock( Cloudflare::class);
-		$cloudflare->shouldReceive('purge_cloudflare')->andReturn( $wp_error );
-
-		$_GET['_wpnonce'] = '';
-		Functions\when( 'wp_verify_nonce' )->justReturn( true );
-		Functions\when( 'sanitize_key' )->justReturn( '' );
-		Functions\when( 'current_user_can' )->justReturn( true );
-		Functions\when( 'get_current_user_id' )->justReturn( 1 );
-
-		$cf_purge_result = [
-			'result'  => 'error',
-			// translators: %s = CloudFare API return message.
-			'message' => sprintf( __( '<strong>WP Rocket:</strong> %s', 'rocket' ), 'Error!' ),
-		];
-
-		Functions\expect( 'set_transient' )
-			->once()
-			->with('1_cloudflare_purge_result', $cf_purge_result );
-
-		$cloudflare_subscriber = new Subscriber( $cloudflare, $mocks['options_data'], $mocks['options'] );
-
-		$cloudflare_subscriber->purge_cache_no_die();
-	}
-
-	/**
-	 * Test should Purge Cloudflare and return success.
-	 */
-	public function testShouldPurgeWithSuccess() {
-		$mocks = $this->getConstructorMocks( 1, '', '', '' );
-
-		$cloudflare = Mockery::mock( Cloudflare::class);
-		$cloudflare->shouldReceive('purge_cloudflare')->andReturn( true );
-
-		$_GET['_wpnonce'] = '';
-		Functions\when( 'wp_verify_nonce' )->justReturn( true );
-		Functions\when( 'sanitize_key' )->justReturn( '' );
-		Functions\when( 'current_user_can' )->justReturn( true );
-		Functions\when( 'get_current_user_id' )->justReturn( 1 );
-
-		$cf_purge_result = [
-			'result'  => 'success',
-			'message' => __( '<strong>WP Rocket:</strong> Cloudflare cache successfully purged.', 'rocket' ),
-		];
-
-		Functions\expect( 'set_transient' )
-			->once()
-			->with('1_cloudflare_purge_result', $cf_purge_result );
-
-		$cloudflare_subscriber = new Subscriber( $cloudflare, $mocks['options_data'], $mocks['options'] );
-
-		$cloudflare_subscriber->purge_cache_no_die();
-	}
-
-	/**
-	 * Get the mocks required by Cloudflare’s constructor.
-	 *
-	 * @since  1.0
-	 * @author Soponar Cristina
-	 * @access private
-	 *
-	 * @param integer $do_cloudflare      - Value to return for $options->get( 'do_cloudflare' ).
-	 * @param string  $cloudflare_email   - Value to return for $options->get( 'cloudflare_email' ).
-	 * @param string  $cloudflare_api_key - Value to return for $options->get( 'cloudflare_api_key' ).
-	 * @param string  $cloudflare_zone_id - Value to return for $options->get( 'cloudflare_zone_id' ).
-	 * @return array                      - Array of Mocks
-	 */
-	private function getConstructorMocks( $do_cloudflare = 1, $cloudflare_email = '',  $cloudflare_api_key = '', $cloudflare_zone_id = '') {
-		$options      = $this->createMock('WP_Rocket\Admin\Options');
-		$options_data = $this->createMock('WP_Rocket\Admin\Options_Data');
-		$map     = [
-			[
-				'do_cloudflare',
-				'',
-				$do_cloudflare,
-			],
-			[
-				'cloudflare_email',
-				null,
-				$cloudflare_email,
-			],
-			[
-				'cloudflare_api_key',
-				null,
-				$cloudflare_api_key,
-			],
-			[
-				'cloudflare_zone_id',
-				null,
-				$cloudflare_zone_id,
-			],
-		];
-		$options_data->method('get')->will( $this->returnValueMap( $map ) );
-
-		$mocks = [
-			'options_data' => $options_data,
-			'options'      => $options,
-		];
-
-		return $mocks;
-	}
-}
